tests for fastapi
