+++ {{cookiecutter.__src_folder_name}}/tests/test_playwright.py
@@ -1,87 +1,11 @@
-import os
-import pathlib
 import re
 
 import pytest
-{% if cookiecutter.project_backend == "flask" %}
-from flask import url_for
-{% endif %}
-{% if cookiecutter.project_backend == "django" %}
-from django.core.management import call_command
-{% endif %}
 from playwright.sync_api import Page, expect
 
-{% if cookiecutter.project_backend == "flask" %}
-from flaskapp import create_app, db, seeder
-{% endif %}
-
-{% if cookiecutter.project_backend == "django" %}
-@pytest.fixture(scope="session")
-def django_db_setup(django_db_setup, django_db_blocker):
-    with django_db_blocker.unblock():
-        call_command("loaddata", "seed_data.json")
-{% endif %}
-
-
-{% if cookiecutter.project_backend == "flask" %}
-@pytest.fixture(scope="session")
-def app():
-    """Session-wide test `Flask` application."""
-    config_override = {
-        "TESTING": True,
-        "SQLALCHEMY_DATABASE_URI": os.environ.get(
-            "TEST_DATABASE_URL", "postgresql://postgres:postgres@localhost:5432/postgres"
-        ),
-    }
-    app = create_app(config_override)
-
-    with app.app_context():
-        engines = db.engines
-        db.create_all()
-        seeder.seed_data(db, pathlib.Path(__file__).parent.parent / "seed_data.json")
-
-    engine_cleanup = []
-
-    for key, engine in engines.items():
-        connection = engine.connect()
-        transaction = connection.begin()
-        engines[key] = connection
-        engine_cleanup.append((key, engine, connection, transaction))
-
-    yield app
-
-    for key, engine, connection, transaction in engine_cleanup:
-        transaction.rollback()
-        connection.close()
-        engines[key] = engine
-{% endif %}
-
-
-@pytest.fixture(scope="session")
-def mock_functions_env():
-    {% if cookiecutter.project_backend == "django" %}
-    os.environ['DJANGO_ALLOW_ASYNC_UNSAFE'] = "true"
-    {% else %}
-    pass
-    {% endif %}
-
-
-def get_index_url(live_server):
-    {% if cookiecutter.project_backend == "django" %}
-    return live_server.url
-    {% endif %}
-    {% if cookiecutter.project_backend == "flask" %}
-    return url_for('pages.index', _external=True)
-    {% endif %}
-
-
-def expect_header_url_test(page:Page, url:str, name: str, re_str: str) -> None:
-    """Helper function to test that a link has a certain attribute"""
-
-
-def test_home(mock_functions_env, page: Page, live_server):
+def test_home(mock_functions_env, page: Page, live_server_url: str):
     """Test that the home page loads"""
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
     expect(page).to_have_title("ReleCloud - Expand your horizons")
     page.close()
 
@@ -94,9 +18,9 @@ def test_home(mock_functions_env, page: Page, live_server):
         ("About", "about"),
     )
 )
-def test_header_has_request_info(mock_functions_env, page: Page, live_server, page_title, page_url):
+def test_header_has_request_info(mock_functions_env, page: Page, live_server_url: str, page_title, page_url):
     """Test that the header loads with links"""
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
     header = page.locator("nav")
 
     # Request Info
@@ -105,16 +29,16 @@ def test_header_has_request_info(mock_functions_env, page: Page, live_server, pa
     page.close()
 
 
-def test_request_information(mock_functions_env, page: Page, live_server):
+def test_request_information(mock_functions_env, page: Page, live_server_url: str):
     """Test that the request info form page loads"""
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
     page.get_by_role("link", name="Request Information").click()
     expect(page).to_have_title("ReleCloud - Request information")
     page.close()
 
 
-def test_destinations(mock_functions_env, page: Page, live_server):
-    page.goto(get_index_url(live_server))
+def test_destinations(mock_functions_env, page: Page, live_server_url: str):
+    page.goto(live_server_url)
     page.get_by_role("link", name="Destinations").click()
     expect(page).to_have_title("ReleCloud - Destinations")
 
@@ -146,13 +70,13 @@ cruises = (
 def test_destination_options(
     page: Page,
     mock_functions_env,
-    live_server,
+    live_server_url: str,
     destination,
     ):
     """Test that the destinations page loads with seeded data"""
 
     # Create a destination
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
 
     page.get_by_role("link", name="Destinations").click()
     expect(page).to_have_title("ReleCloud - Destinations")
@@ -165,10 +89,10 @@ def test_destination_options(
 def test_destination_options_have_cruises(
     page: Page,
     mock_functions_env,
-    live_server,
+    live_server_url: str,
     destination
 ):
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
     page.get_by_role("link", name="Destinations").click()
     page.get_by_role("link", name=destination).click()
     expect(page).to_have_url(re.compile(r".*destination/\d+", re.IGNORECASE))
@@ -181,9 +105,9 @@ def test_destination_options_have_cruises(
     page.close()
 
 
-def test_about(mock_functions_env, page: Page, live_server):
+def test_about(mock_functions_env, page: Page, live_server_url:str):
     """Test that the request info form page loads"""
-    page.goto(get_index_url(live_server))
+    page.goto(live_server_url)
     page.get_by_role("link", name="About").click()
     expect(page.locator("#page-title")).to_have_text(re.compile(r".*about.*", re.IGNORECASE))
     page.close()
